requirements.txt
audioread==2.1.8
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-cors-headers==3.1.1
djangorestframework==3.10.3
gunicorn==20.0.0
idna==2.8
pyacoustid==1.1.7
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
